requirements.txt
# This project uses uv for package management
# Install with: uv pip install -r requirements.txt
# Update with: uv pip freeze > requirements.txt

# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
email-validator==2.1.0.post1

# WebSockets
websockets==11.0.3

# Monitoring and metrics
psutil==5.9.6

# Logging
# Using standard Python logging module

# Production
gunicorn==21.2.0
uvloop==0.19.0; sys_platform != 'win32'
httptools==0.6.1
